feat(entry): add entry reasoning plugin
# Conflicts:
#	src/entry_reasoning_pipeline.py
